# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zonda API
  version: 0.1.0
paths:
  /sign-up:
    post:
      tags: [authentication]
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              required: [email, password]
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                required: [access_token, refresh_token]
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /sign-in:
    post:
      tags: [authentication]
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              required: [email, password]
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                required: [access_token, refresh_token]
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/error"
  /health:
    get:
      tags: [monitoring]
      operationId: checkHealth
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  revision:
                    type: string
                required: [revision]
components:
  schemas:
    error:
      type: object
      properties:
        code:
          type: number
        message:
          type: string
      required: [code, message]
tags:
  - name: authentication
  - name: monitoring